+++ webhooks/virtualmachinesetresourcepolicy/v1alpha1/mutation/virtualmachinesetresourcepolicy_mutator.go
@@ -52,7 +52,6 @@ type mutator struct {
 	converter runtime.UnstructuredConverter
 }
 
-//nolint
 func (m mutator) Mutate(ctx *context.WebhookRequestContext) admission.Response {
 	vmRP, err := m.vmRPFromUnstructured(ctx.Obj)
 	if err != nil {
